rag/Multimodal_RAG_Project-main/backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.13.2
fastapi==0.135.3
filelock==3.25.2
fsspec==2026.3.0
ftfy==6.3.1
greenlet==3.4.0
groq==1.1.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
lxml==6.0.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
open_clip_torch==3.3.0
packaging==26.0
pillow==12.2.0
pluggy==1.6.0
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
Pygments==2.20.0
pypdf==6.10.0
pytest==9.0.3
pytest-asyncio==1.3.0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.26
python-pptx==1.0.2
PyYAML==6.0.3
rank-bm25==0.2.2
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.0
setuptools==81.0.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
structlog==25.5.0
sympy==1.14.0
tavily-python==0.3.3
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.26
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
xlsxwriter==3.2.9
